Add basic Jest tests and scripts
