+++ backend/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node server.js",
     "dev": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "keywords": [
     "bvmw",
@@ -26,7 +26,8 @@
     "mysql2": "^3.4.0"
   },
   "devDependencies": {
-    "nodemon": "^2.0.22"
+    "jest": "^30.0.4",
+    "nodemon": "^2.0.22",
+    "supertest": "^7.1.3"
   }
 }
-
